Finished FDE ratcheting (XEX) model, basic runtime testing

+++ fde-c/include/kdf.h
@@ -3,6 +3,6 @@
 
 #include "common.h"
 
-int expand_master_key(const byte* master_key, byte* key_0, byte* key_F, byte* key_G, size_t key_length);
+int expand_master_key(const byte* master_key, byte* key_0, byte* key_F, byte* key_G, uint32_t key_length);
 
 #endif
