+++ fde-c/include/primitives.h
@@ -3,8 +3,8 @@
 
 #include "common.h"
 
-void xor_bytes(const byte* a, const byte* b, byte* out, size_t length);
-int aes_block_encrypt(const byte* key, const byte* block, byte* out, size_t key_length);
-int aes_block_decrypt(const byte* key, const byte* block, byte* out, size_t key_length);
+void xor_bytes(const byte* a, const byte* b, byte* out, uint32_t length);
+int aes_block_encrypt(const byte* key, const byte* block, byte* out, uint32_t key_length);
+int aes_block_decrypt(const byte* key, const byte* block, byte* out, uint32_t key_length);
 
 #endif
