+++ fde-c/include/test_utils.h
@@ -3,7 +3,7 @@
 
 #include "common.h"
 
-void hex_to_bytes(const char* hex, byte* out, size_t out_len);
-void bytes_to_hex(const byte* bytes, char* out, size_t out_len);
+void hex_to_bytes(const char* hex, byte* out, uint32_t out_len);
+void bytes_to_hex(const byte* bytes, char* out, uint32_t out_len);
 
 #endif
