+++ fde-c/src/test_utils.c
@@ -2,14 +2,14 @@
 
 #include <stdio.h>
 
-void hex_to_bytes(const char* hex, byte* out, size_t out_len) {
-	for (size_t i = 0; i < out_len; i++) {
+void hex_to_bytes(const char* hex, byte* out, uint32_t out_len) {
+	for (uint32_t i = 0; i < out_len; i++) {
 		sscanf(hex + 2 * i, "%2hhx", &out[i]);
 	}
 }
 
-void bytes_to_hex(const byte* bytes, char* out, size_t out_len) {
-	for (size_t i = 0; i < out_len; i++) {
+void bytes_to_hex(const byte* bytes, char* out, uint32_t out_len) {
+	for (uint32_t i = 0; i < out_len; i++) {
 		sprintf(out + 2 * i, "%02x", bytes[i]);
 	}
 }
