+++ fde-c/src/xex.c
@@ -1,6 +1,6 @@
 #include "xex.h"
 
-int xex_encrypt_block(const byte* key, const byte* tweak, const byte* block, byte* out, size_t length) {
+int xex_encrypt_block(const byte* key, const byte* tweak, const byte* block, byte* out, uint32_t length) {
 	byte masked[length];
 	byte core[length];
 
@@ -14,7 +14,7 @@ int xex_encrypt_block(const byte* key, const byte* tweak, const byte* block, byt
 	return 0;
 }
 
-int xex_decrypt_block(const byte* key, const byte* tweak, const byte* block, byte* out, size_t length) {
+int xex_decrypt_block(const byte* key, const byte* tweak, const byte* block, byte* out, uint32_t length) {
 	byte masked[length];
 	byte core[length];
 
@@ -28,7 +28,7 @@ int xex_decrypt_block(const byte* key, const byte* tweak, const byte* block, byt
 	return 0;
 }
 
-int update_state(const byte* key, const byte* tweak, const byte* block, const byte* key_F, const byte* key_G, byte* next_key, byte* next_tweak, size_t length) {
+int update_state(const byte* key, const byte* tweak, const byte* block, const byte* key_F, const byte* key_G, byte* next_key, byte* next_tweak, uint32_t length) {
 	byte masked[length];
 	int result;
